+++ Student_Cuda/src/cpp/core/04_Montecarlo/01_object/host/Montecarlo.h
@@ -24,10 +24,9 @@ class Montecarlo{
 
     private:
 	dim3 dg, db;
-
 	int nbFlechette;
 
-	int M;
+	float M;
 	float a, b;
 
 	int nbFlechetteDessous;
